+++ src/mare/CMake.cpp
@@ -418,7 +418,7 @@ bool CMake::writeProject(const String& targetName, Project& project)
       for(const List<String>::Node* i = target.cppCompiler.getFirst(); i; i = i->getNext())
       {
         const String& word = i->data;
-        unsigned int sep;
+        size_t sep;
         if(word.find('=', sep))
           fileWrite(String("set(ENV{") + word.substr(0, sep) +  "} \"" + word.substr(sep + 1) + "\")\n");
         else
@@ -442,7 +442,7 @@ bool CMake::writeProject(const String& targetName, Project& project)
       for(const List<String>::Node* i = target.cCompiler.getFirst(); i; i = i->getNext())
       {
         const String& word = i->data;
-        unsigned int sep;
+        size_t sep;
         if(word.find('=', sep))
           fileWrite(String("set(ENV{") + word.substr(0, sep) +  "} \"" + word.substr(sep + 1) + "\")\n");
         else
